5-asynchronous/cdk/requirements.txt
aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-lambda-destinations
